+++ NPS Browser/models/Bookmark.swift
@@ -14,14 +14,14 @@ class Bookmark: NSObject {
     @objc dynamic var type                      : String
     @objc dynamic var zrif                      : String?
     @objc dynamic var pkg_direct_link           : URL?
-    @objc dynamic var sha256                    : String
-    init(name: String, title_id: String, type: String, zrif: String?, pkg_direct_link: URL?, sha256: String) {
+    @objc dynamic var uuid                      : UUID
+    init(name: String, title_id: String, type: String, zrif: String?, pkg_direct_link: URL?, uuid: UUID) {
         self.name                       = name
         self.title_id                   = title_id
         self.type                       = type
         self.zrif                       = zrif
         self.pkg_direct_link            = pkg_direct_link
-        self.sha256                     = sha256
+        self.uuid                       = uuid
         super.init()
     }
 }